drivers: nl80211: correct scan flags type
Aggregate scan flags in a u32 instead of u8 to match the parameter.  Should
have no effect right now as the variable is used by value.

BUG=none
TEST=wpa_cli scan works

Reviewed-on: http://gerrit.chromium.org/gerrit/9722
Commit-Ready: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>

+++ src/drivers/driver_nl80211.c
@@ -2335,7 +2335,7 @@ static int wpa_driver_nl80211_scan(void *priv,
 	int ret = 0, timeout;
 	struct nl_msg *msg, *ssids, *freqs;
 	size_t i;
-	u8 flags;
+	u32 flags;
 
 	msg = nlmsg_alloc();
 	ssids = nlmsg_alloc();
